Create initial Processing QC
Processing script for the inital data for the Phenology group work. Does require two libraries for functions used in the script. Assumes the formatting is in an AmeriFlux style format; does some QA checks and creates the output file to be put into REddyProc. Note this is NOT the broader flux processing script; assumes most of the QC has been performed prior to this point.